+++ bin/countly.install.sh
@@ -14,7 +14,6 @@ echo "
 / /___/ /_/ / /_/ / / / / /_/ / /_/ /
 \____/\____/\__,_/_/ /_/\__/_/\__, /
               http://count.ly/____/
-
 "
 
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
@@ -22,14 +21,15 @@ DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 #update package index
 apt-get update
 
-apt-get -y install python-software-properties
+apt-get -y install python-software-properties wget g++
 
 if !(command -v apt-add-repository >/dev/null) then
     apt-get -y install software-properties-common
 fi
 
 #add node.js repo
-echo | apt-add-repository ppa:chris-lea/node.js
+#echo | apt-add-repository ppa:chris-lea/node.js
+wget -qO- https://deb.nodesource.com/setup | bash -
 
 #add mongodb repo
 echo "deb http://downloads-distro.mongodb.org/repo/ubuntu-upstart dist 10gen" > /etc/apt/sources.list.d/mongodb-10gen-countly.list
@@ -48,7 +48,10 @@ apt-get -y --force-yes install nodejs || (echo "Failed to install nodejs." ; exi
 apt-get -y --force-yes install mongodb-org || (echo "Failed to install mongodb." ; exit)
 
 #install supervisor
-apt-get -y install supervisor || (echo "Failed to install supervisor." ; exit)
+if [ "$1" != "docker" ]
+then
+	apt-get -y install supervisor || (echo "Failed to install supervisor." ; exit)
+fi
 
 #install imagemagick
 apt-get -y install imagemagick
@@ -56,40 +59,66 @@ apt-get -y install imagemagick
 #install sendmail
 apt-get -y install sendmail
 
+#apt-get -y install build-essential || (echo "Failed to install build-essential." ; exit)
+
 #install iptables
-apt-get -y install iptables
+if [ "$1" != "docker" ]
+then
+	DEBIAN_FRONTEND=noninteractive apt-get -y install iptables-persistent
 
-apt-get -y install build-essential || (echo "Failed to install build-essential." ; exit)
+	#drop packages coming from 0/0 going through mongodb port
+	#allow those coming from localhost
+	iptables -A INPUT -m state --state NEW -p tcp --destination-port 27019 -s localhost -j ACCEPT
+	iptables -A INPUT -m state --state NEW -p tcp --destination-port 27019 -s 0/0 -j DROP
 
-#drop packages coming from 0/0 going through mongodb port
-#allow those coming from localhost
-iptables -A INPUT -m state --state NEW -p tcp --destination-port 27019 -s localhost -j ACCEPT
-iptables -A INPUT -m state --state NEW -p tcp --destination-port 27019 -s 0/0 -j DROP
+	#install iptables-persistent
+	apt-get -y install iptables-persistent
+fi
 
-#install iptables-persistent
-apt-get -y install iptables-persistent
-
-#install time module for node
-( cd $DIR/../api ; npm install time )
+#install grunt & npm modules
+( cd $DIR/.. ; npm install -g grunt-cli --unsafe-perm ; npm install )
 
 #configure and start nginx
 cp /etc/nginx/sites-enabled/default $DIR/config/nginx.default.backup
 cp $DIR/config/nginx.server.conf /etc/nginx/sites-enabled/default
-/etc/init.d/nginx restart
+cp $DIR/config/nginx.conf /etc/nginx/nginx.conf
+if [ "$1" != "docker" ]
+then
+	/etc/init.d/nginx restart
+fi
 
-cp $DIR/../frontend/express/public/javascripts/countly/countly.config.sample.js $DIR/../frontend/express/public/javascripts/countly/countly.config.js
+cp -n $DIR/../frontend/express/public/javascripts/countly/countly.config.sample.js $DIR/../frontend/express/public/javascripts/countly/countly.config.js
 
 #kill existing supervisor process
-pkill -SIGTERM supervisord
+if [ "$1" != "docker" ]
+then
+	pkill -SIGTERM supervisord
 
-#create supervisor upstart script
-(cat $DIR/config/countly-supervisor.conf ; echo "exec /usr/bin/supervisord --nodaemon --configuration $DIR/config/supervisord.conf") > /etc/init/countly-supervisor.conf
+	#create supervisor upstart script
+	(cat $DIR/config/countly-supervisor.conf ; echo "exec /usr/bin/supervisord --nodaemon --configuration $DIR/config/supervisord.conf") > /etc/init/countly-supervisor.conf
+fi
+
+#respawning mongod on crash
+echo "respawn" >> /etc/init/mongod.conf
 
 #create api configuration file from sample
-cp $DIR/../api/config.sample.js $DIR/../api/config.js
+cp -n $DIR/../api/config.sample.js $DIR/../api/config.js
 
 #create app configuration file from sample
-cp $DIR/../frontend/express/config.sample.js $DIR/../frontend/express/config.js
+cp -n $DIR/../frontend/express/config.sample.js $DIR/../frontend/express/config.js
+
+if [ ! -f $DIR/../plugins/plugins.json ]; then
+	cp $DIR/../plugins/plugins.default.json $DIR/../plugins/plugins.json
+fi
+
+#install plugins
+bash $DIR/scripts/countly.install.plugins.sh
+
+#compile scripts for production
+cd $DIR && grunt dist-all
 
 #finally start countly api and dashboard
-start countly-supervisor
+if [ "$1" != "docker" ]
+then
+	start countly-supervisor
+fi
